Delete TaskList, move to ToDoList
